lib_libfflush, up_assert, serial: Add to flush stdout buffer in assert
If board is crash(abort), we use only lldbg to printing logs.
Becasue OS operation, including uart driver, could be incorrect.
So, In this time, printf and dbg operation is stopped, these do not complete output.
And logs are remaining in the buffer. so, app last log is cutted off.

Therefore, Add to flush the logs remaining in the buffer after outputting the assertion logs.

But, This flush operation could has crush. we need to prevent in this case.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>