drivers/serial, arch: Change uart_xmitchars return type to uint16_t
When system fault is occured, The system is in the irq locking state, The logs that have not
yet been output remain in the uart buffer and are not output.
So, For debugging, it is necessary to flush the log uart buffer before outputting the fault log.

For flushing transmit buffe, we need to know xmit buffer is empty or not.
Therefore, Add to so that uart_xmitchars function returns amount sent.
the return value is 0, xmit buffer is empty.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>